Add stub file for collection.py
I would like to note that I am currently returning the Any type for methods
that return a document "_id". This is due to the fact that while the default
behavior is to return a bson.ObjectID, a user could manually set the _id
field and thus the return value could be anything.